requirements.txt
ftfy==6.3.1
google-ai-generativelanguage==0.6.10
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3

pickle
numpy
open_clip_torch==2.30.0


tensorboard==2.18.0
tensorboard-data-server==0.7.2

tokenizers==0.20.2

torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1

tqdm

transformers==4.46.1
